Move `getDocsUrl` out from `server/routing.js` to `server/utils.js`
**Why?**
Because `routing.js` is exporting all router RegEx's, and the
`getDocsUrl` suffices more as a util
