+++ v1/lib/server/routing.js
@@ -4,8 +4,7 @@
  * This source code is licensed under the MIT license found in the
  * LICENSE file in the root directory of this source tree.
  */
-
-const DOCS_URL = 'docs';
+import {DOCS_URL} from './utils';
 
 function blog(baseUrl) {
   return new RegExp(`^${baseUrl}blog/.*html$`);
@@ -38,10 +37,6 @@ function sitemap(baseUrl) {
   return new RegExp(`^${baseUrl}sitemap.xml$`);
 }
 
-function getDocsUrl(docsUrl) {
-  return docsUrl || DOCS_URL;
-}
-
 module.exports = {
   blog,
   docs,
@@ -50,6 +45,4 @@ module.exports = {
   page,
   noExtension,
   sitemap,
-  getDocsUrl,
-  DOCS_URL,
 };
