+++ v1/website/pages/en/help.js
@@ -12,7 +12,7 @@ const Container = CompLibrary.Container;
 const GridBlock = CompLibrary.GridBlock;
 const siteConfig = require(`${process.cwd()}/siteConfig.js`);
 const translate = require('../../server/translate.js').translate;
-const {getDocsUrl} = require('../../server/routing.js');
+const {getDocsUrl} = require('../../server/utils.js');
 
 class Help extends React.Component {
   render() {
@@ -21,9 +21,7 @@ class Help extends React.Component {
         title: <translate>Browse the docs</translate>,
         content: `Learn more about Docusaurus using the [official documentation](${
           siteConfig.baseUrl
-        }${getDocsUrl(siteConfig.docsUrl)}/${
-          this.props.language
-        }/installation).`,
+        }${getDocsUrl(siteConfig)}/${this.props.language}/installation).`,
       },
       {
         title: <translate>Discord</translate>,
